refactor: split P2P handlers into focused modules (Q-DEVNET-12)
