+++ clients/go/node/p2p/service.go
@@ -105,66 +105,6 @@ func NewService(cfg ServiceConfig) (*Service, error) {
 	}, nil
 }
 
-func (s *Service) Start(ctx context.Context) error {
-	if s == nil {
-		return errors.New("nil service")
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if s.listener != nil {
-		return errors.New("service already started")
-	}
-	listener, err := net.Listen("tcp", s.cfg.BindAddr)
-	if err != nil {
-		return err
-	}
-	s.listener = listener
-	s.ctx, s.cancel = context.WithCancel(ctx)
-
-	go s.acceptLoop()
-	for _, peerAddr := range s.cfg.BootstrapPeers {
-		peerAddr = strings.TrimSpace(peerAddr)
-		if peerAddr == "" {
-			continue
-		}
-		go s.dialPeer(peerAddr)
-	}
-	return nil
-}
-
-func (s *Service) Close() error {
-	if s == nil {
-		return nil
-	}
-	if s.cancel != nil {
-		s.cancel()
-	}
-	if s.listener != nil {
-		_ = s.listener.Close()
-	}
-	s.peersMu.RLock()
-	peers := make([]*peer, 0, len(s.peers))
-	for _, current := range s.peers {
-		peers = append(peers, current)
-	}
-	s.peersMu.RUnlock()
-	for _, current := range peers {
-		_ = current.conn.Close()
-	}
-	return nil
-}
-
-func (s *Service) Addr() string {
-	if s == nil {
-		return ""
-	}
-	if s.listener != nil {
-		return s.listener.Addr().String()
-	}
-	return s.cfg.BindAddr
-}
-
 func (s *Service) AnnounceBlock(blockBytes []byte) error {
 	if s == nil {
 		return errors.New("nil service")
@@ -199,108 +139,4 @@ func (s *Service) AnnounceTx(txBytes []byte) error {
 		return nil
 	}
 	return s.broadcastInventory(nil, []InventoryVector{{Type: MSG_TX, Hash: txid}})
-}
-
-func (s *Service) acceptLoop() {
-	for {
-		conn, err := s.listener.Accept()
-		if err != nil {
-			if s.ctx != nil && s.ctx.Err() != nil {
-				return
-			}
-			if errors.Is(err, net.ErrClosed) {
-				return
-			}
-			continue
-		}
-		go s.handleConn(conn)
-	}
-}
-
-func (s *Service) dialPeer(addr string) {
-	dialer := &net.Dialer{Timeout: s.cfg.PeerRuntimeConfig.HandshakeTimeout}
-	conn, err := dialer.DialContext(s.ctx, "tcp", addr)
-	if err != nil {
-		return
-	}
-	s.handleConn(conn)
-}
-
-func (s *Service) handleConn(conn net.Conn) {
-	defer func() {
-		if conn != nil {
-			_ = conn.Close()
-		}
-	}()
-
-	localVersion, err := s.localVersion()
-	if err != nil {
-		return
-	}
-	state, err := performHandshake(
-		s.ctx,
-		conn,
-		s.cfg.PeerRuntimeConfig,
-		localVersion,
-		s.cfg.SyncConfig.ChainID,
-		s.cfg.GenesisHash,
-	)
-	if err != nil {
-		return
-	}
-
-	current := &peer{
-		conn:    conn,
-		service: s,
-		state:   state,
-	}
-	if err := s.registerPeer(current); err != nil {
-		return
-	}
-	defer s.unregisterPeer(current.addr())
-
-	s.cfg.SyncEngine.RecordBestKnownHeight(state.RemoteVersion.BestHeight)
-	if err := s.requestBlocksIfBehind(current); err != nil {
-		current.setLastError(err.Error())
-		return
-	}
-	if err := current.run(s.ctx); err != nil && s.ctx.Err() == nil {
-		current.setLastError(err.Error())
-		return
-	}
-}
-
-func (s *Service) registerPeer(p *peer) error {
-	if err := s.cfg.PeerManager.AddPeer(&p.state); err != nil {
-		return err
-	}
-	s.peersMu.Lock()
-	defer s.peersMu.Unlock()
-	s.peers[p.addr()] = p
-	return nil
-}
-
-func (s *Service) unregisterPeer(addr string) {
-	s.peersMu.Lock()
-	delete(s.peers, addr)
-	s.peersMu.Unlock()
-	s.cfg.PeerManager.RemovePeer(addr)
-}
-
-func (s *Service) localVersion() (node.VersionPayloadV1, error) {
-	bestHeight, _, ok, err := s.cfg.BlockStore.Tip()
-	if err != nil {
-		return node.VersionPayloadV1{}, err
-	}
-	if !ok {
-		bestHeight = 0
-	}
-	return node.VersionPayloadV1{
-		Magic:           ProtocolMagic,
-		ProtocolVersion: ProtocolVersion,
-		ChainID:         s.cfg.SyncConfig.ChainID,
-		GenesisHash:     s.cfg.GenesisHash,
-		UserAgent:       s.cfg.UserAgent,
-		BestHeight:      bestHeight,
-	}, nil
 }
